+++ changelog.md
@@ -1,4 +1,5 @@
 # Change Log
+
 Changes
 
 ## [0.0.8] - 2023-01-02
@@ -6,9 +7,9 @@ Changes
 ### Added
 
 - Created a client `create react-app client`
-- Installe react route `react-router-dom@6.4.3` 
+- Install react route `react-router-dom@6.4.3`
+- Installed `cypress@10.11.0`
 
 ### Changed
 
 - Setup Basic Routes
-